refactor(git): move TagInfo into its own src/git/tag.rs module
Mirrors the existing branch.rs/commit.rs/diff.rs split-by-concept
convention and gives room for future tag-specific fields (tagger,
message, annotated flag) without bloating branch.rs.

Addresses Hunk review feedback.
