+++ src/git/mod.rs
@@ -7,8 +7,9 @@ pub mod extensions;
 pub mod graph;
 pub mod operations;
 pub mod repository;
+pub mod tag;
 
-pub use branch::{BranchInfo, TagInfo};
+pub use branch::BranchInfo;
 pub use commit::CommitInfo;
 pub use diff::{
     CommitDiffInfo, DiffHunkContent, DiffLineContent, DiffLineOrigin, FileChangeKind,
@@ -17,3 +18,4 @@ pub use diff::{
 pub use extensions::configure_git_extensions;
 pub use graph::build_graph;
 pub use repository::{GitRepository, StageState, WorkingTreeStatus};
+pub use tag::TagInfo;
